Rearrange resource due to GCC compiler bug

+++ include/md/dma.h
@@ -3,11 +3,11 @@
 #include <stddef.h>
 
 struct DmaQueueItem {
-	uint32_t sourceAddress;
+	uintptr_t sourceAddress;
 	uint16_t size;
 	uint32_t destinationWord;
 };
 
 struct DmaQueueItem dmaQueue[ 15 ];
 
-void mdDmaEnqueue( uint32_t source, uint16_t size, uint32_t destinationWord );
+void mdDmaEnqueue( uintptr_t source, uint16_t size, uint32_t destinationWord );
